Clean Code in components
